requirements.txt
-r requirements/base.txt

# this file is only heroku specific
gunicorn==20.1.0
django-heroku==0.3.1
